Frog Input Handler
The two frogs now has independent buttons in their side of the screen.
I haven't make the enemies work, not even make them appear in the
screen.
I'm trying to approach to the solution by working with queues, but still
haven't seen results. On next commit, remember to check the nullability